Log the segments
We are currently passing `Collection` into `entity_set_by_name` which is returning `nil`. Downstream, that ends up calling `member_type` on that `nil` and errors. By adding the logging, I am trying to understand what the segments actually are in this case.

We already have some logging that shows us the available sets like:

`EntitySet "Collection" not found. Sets are ["directoryObjects", "devices", "groups", "directoryRoles", "directoryRoleTemplates", "organization", "subscribedSkus", "users", "drives"]`

Hopefully, after this, we'll be able to see something else in segments that we should be grabbing in this case.